refactor: move conteúdo geral, como erros, pra crate `comum`
