+++ comum/src/erros/mod.rs
