+++ comum/src/erros/resultado_de_dominio.rs
@@ -1,3 +1,3 @@
-use crate::utils::erros::erro_de_dominio::ErroDeDominio;
+use crate::erros::erro_de_dominio::ErroDeDominio;
 
 pub type ResultadoDominio<T> = Result<T, ErroDeDominio>;
