+++ comum/src/erros/tipo_erro_de_dominio.rs
