+++ comum/src/extensions.rs
@@ -1,0 +1,2 @@
+mod actix;
+mod inertia;
