+++ comum/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod erros;
+mod extensions;
+pub mod sqlx;
