+++ comum/src/sqlx/nullableu8.rs
@@ -1,8 +1,6 @@
 use std::ops::Deref;
 
-use pbkdf2::hmac::digest::typenum::Cmp;
 use serde::{Deserialize, Serialize};
-use sqlx::prelude::Type;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct NullableU8(pub Option<u8>);
