+++ src/libs/mod.rs
@@ -1,8 +1,6 @@
 use chrono::{DateTime, Utc};
 
 pub mod actix;
-pub mod inertia;
-pub mod sqlx;
 pub mod validator;
 
 pub type UtcDateTime = DateTime<Utc>;
