--- src/libs/sqlx.rs
@@ -1,2 +1,0 @@
-mod from_row;
-mod from_row_with_prefix;
